wpt_babel/content-security-policy/script-src/scripthash-ignore-unsafeinline.sub.html

<!DOCTYPE html>
<html>
 <head>
  <!-- Programmatically converted from a WebKit Reftest, please forgive resulting idiosyncracies.-->
  <meta content="script-src 'self' 'unsafe-inline' 'sha256-3iveTSiUbmzN7COYvdDwyaXXzJ3SrjKlTaOvQ/GdRpo=' 'sha256-EgE/bwVJ+ZLL9F5hNjDqD4C7nlFFrdDaKeNIJ2cUem4=' 'sha256-lxHfHAe5I15v8qaArcZ5WiKmLU4CjV+3tJeQUqSIWBk='; connect-src 'self';" http-equiv="Content-Security-Policy"/>
  <title>
   scripthash-ignore-unsafeinline
  </title>
  <script src="/resources/testharness.js">
  </script>
  <script src="/resources/testharnessreport.js">
  </script>
  <script src="../support/logTest.sub.js?logs=[]">
  </script>
  <script>
   "use strict";

window.addEventListener('securitypolicyviolation', function (e) {
  alert_assert("Fail");
});
  </script>
  <script>
   "use strict";

var t_alert = async_test('Expecting alerts: ["PASS (1/1)"]');
var expected_alerts = ["PASS (1/1)"];
function alert_assert(msg) {
  t_alert.step(function () {
    if (msg.match(/^FAIL/i)) {
      assert_unreached(msg);
      t_alert.done();
    }
    for (var i = 0; i < expected_alerts.length; i++) {
      if (expected_alerts[i] == msg) {
        assert_equals(expected_alerts[i], msg);
        expected_alerts.splice(i, 1);
        if (expected_alerts.length == 0) {
          t_alert.done();
        }
        return;
      }
    }
    assert_unreached('unexpected alert: ' + msg);
    t_log.done();
  });
}
  </script>
  <script>
   "use strict";

alert_assert('PASS (1/1)');
  </script>
  <script>
   "use strict";

alert_assert('FAIL (1/1)');
  </script>
 </head>
 <body>
  <p>
   This tests that a valid hash value disables inline JavaScript, even if 'unsafe-inline' is present.
  </p>
  <div id="log">
  </div>
 </body>
</html>
